MieScattering

Documentation for MieScattering.

MieScattering.bhcoat

MieScattering.bhmie

MieScattering.bhcoat — Method

bhcoat([T=Float64,], x_core, x_mantle, m_core, m_mantle; Nθ = 181, tolerance = 1e-8)

Inputs:

T: Type used for calculation. All real numbers will be stored as T, while all complex numbers will be stored as Complex{T}.

x_core: Size parameter of the core. Defined as $\frac{2\pi r}{\lambda}$

x_mantle: Size parameter of the coated sphere. x_mantle >= x_core should hold.

m_core: Refractive index of the core, relative to the host medium.

m_mantle: Refractive index of the mantle, relative to the host medium.

Keyword arguments:

Nθ: Number of scattering angles to calculate. Default is 181.

tolerance: Error tolerance. Default is 1e-8.

Scattering information is outputed as a named tuple, including the following fields:

q_ext: Extinction efficiency. Defined as $C_{\mathrm{ext}}/\pi r^2$, where $C_{\mathrm{ext}}$ is the extinction cross section.

q_sca: Scattering efficiency.

q_abs: Absorption efficiency.

q_back: Backscattering efficiency. Defined as $4(\mathrm{d}C_\mathrm{ext}/\mathrm{d}\Omega)/r^2$.

asymm: Asymmetry factor $\langle\cos(\theta)\rangle$.

S = (S₁, S₂): Amplitude scattering matrix components. See Eq. (3.12) in Bohren and Huffman (1983). Both S₁ and S₂ are vectors containing Nθ values.

F = (F₁₁, F₁₂, F₃₃, F₃₄): Mueller scattering matrix components. See Eq. (3.16) in Bohren and Huffman (1983). All Fᵢⱼ are vectors containing Nθ values.

References:

Bohren, C.F., Huffman, D.R., 1983. Absorption and scattering of light by small particles. John Wiley & Sons.

MieScattering.bhmie — Method

bhmie([T=Float64,], x, m; Nθ=181, nextra=15, custom_nstop=0)

Inputs:

T: Type used for calculation. All real numbers will be stored as T, while all complex numbers will be stored as Complex{T}.

x: Size parameter of the sphere scatterer. Defined as $\frac{2\pi r}{\lambda}$

m: Relative refractive index of the scatterer.

Keyword arguments:

Nθ: Number of scattering angles to calculate. Default is 181.

nextra: Extra terms used for the downward calculation of the d function. Default is 15.

custom_nstop: Custom truncation point. Default is 0, and the empirical formula

\[n_{\mathrm{stop}}=\max(x+4\sqrt[3]{x}+2, |m|x)\]

will be used.

Scattering information is outputed as a named tuple, including the following fields:

q_ext: Extinction efficiency. Defined as $C_{\mathrm{ext}}/\pi r^2$, where $C_{\mathrm{ext}}$ is the extinction cross section.

q_sca: Scattering efficiency.

q_abs: Absorption efficiency.

q_back: Backscattering efficiency. Defined as $4(\mathrm{d}C_\mathrm{ext}/\mathrm{d}\Omega)/r^2$.

asymm: Asymmetry factor $\langle\cos(\theta)\rangle$.

S = (S₁, S₂): Amplitude scattering matrix components. See Eq. (3.12) in Bohren and Huffman (1983). Both S₁ and S₂ are vectors containing Nθ values.

F = (F₁₁, F₁₂, F₃₃, F₃₄): Mueller scattering matrix components. See Eq. (3.16) in Bohren and Huffman (1983). All Fᵢⱼ are vectors containing Nθ values.

References:

Bohren, C.F., Huffman, D.R., 1983. Absorption and scattering of light by small particles. John Wiley & Sons.