Mie

Missing docstring.

Missing docstring for mie_mx. Check Documenter's build log for details.

MieScattering.mie_ab — Function

mie_ab(m, x, nmax)

Computes external field coefficients $a_n$ and $b_n$ based on inputs of m and $x = \pi \, d_p / \lambda$.

Arguments

m: The complex refractive index with the convention $m = n+ik$.

x: The size parameter $x = \pi \, d_p / \lambda$.

Return values

an, bn: Arrays of size $n_max = 2 + x + 4x^{1/3}$

MieScattering.mie_ab_n — Function

mie_abcd(m, x, n::Int)

Calculate n-th order Mie coefficients a,b,c,d

Arguments

m: relative refractive index ($n_\mathrm{material} / n_\mathrm{environment}$)

x: size parameter (wavenumber * radius)

n::Int

MieScattering.mie_pi_tau — Function

mie_pi_tau(mu, nmax)

Calculates $\pi_n$ and $\tau_n$.

This function uses recurrence relations to calculate $\pi_n$ and $\tau_n$, beginning with $\pi_0 = 1$, $\pi_1 = 3 \mu$ (where $mu$ is the cosine of the scattering angle), $\tau_0 = \mu$, and $\tau_1 = 3 cos(2 cos^{-1}(\mu))$:

\[\pi_n = \frac{2n-1}{n-1} \mu \pi_{n-1} - \frac{n}{n-1} \pi_{n-2}\]

\[\tau_n = n \mu \pi_n - (n+1) \pi_{n-1}\]

Arguments

mu : The cosine of the scattering angle.

nmax: The number of elements to compute. Typically, $n_{max} = floor(2+x+4x^{1/3})$, but can be given any integer.

Return values

p, t: The $\pi_n$ and $\tau_n$ arrays, of length nmax.

MieScattering.mie_pi_tau_n — Function

mie_pi_tau(mu, n::Int)

See Sec. 4.3.1 of B&H ref of pi,tau: [D. Deirmendjian, “Electromagnetic Scattering on Spherical Polydispersions"]

MieScattering.mie_scattering — Function

mie_scattering(m, x, [nmax])

Computes Mie efficencies Q and asymmetry parameter g of a single, homogeneous particle. Uses mie_ab to calculate $a_n$ and $b_n$, and then calculates Q via:

\[Q_{ext} = \frac{2}{x^2} \sum \limits_{n=1}^{n_{max}} (2n+1) \: \text{Re} \left\{ a_n + b_n \right\}\]

\[Q_{sca} = \frac{2}{x^2} \sum_{n=1}^{n_{max}} (2n+1) (|a_n|^2 + |b_n|^2)\]

\[Q_{abs} = Q_{ext} - Q_{sca}\]

\[Q_{back} = \frac{1}{x^2} \left| \sum\limits_{n=1}^{n_{max}} (2n+1) (-1)^n (a_n - b_n) \right|^2\]

\[Q_{ratio} = \frac{Q_{back}}{Q_{sca}}\]

\[g = \frac{4}{Q_{sca}x^2} \left[ \sum\limits_{n=1}^{n_{max}} \frac{n(n+2)}{n+1} \text{Re}\left\{a_n a_{n+1}^* + b_n b_{n+1}^*\right\} + \sum\limits_{n=1}^{n_{max}} \frac{2n+1}{n(n+1)} \text{Re}\left\{a_n b_n^*\right\}\right]\]

\[Q_{pr} = Q_{ext} - g Q_{sca}\]

where asterisks denote the complex conjugates.

Arguments

m: The complex refractive index, with the convention m = n+ik.

wavelength: The wavelength of incident light, in nanometers.

diameter: The diameter of the particle, in nanometers.

n_medium: The refractive index of the surrounding medium. This must be positive, nonzero, and real. Any imaginary part will be discarded.

Return values

qext, qsca, qabs, g, qpr, qback, qratio: The Mie efficencies described above.

Examples

For example, compute the Mie efficencies of a particle $300 \, [\mathrm{nm}]$ in diameter with $m = 1.77+0.63i$, illuminated by $\lambda = 375 \, [\mathrm{nm}]$:

julia> mie_scattering(1.77+0.63im, 375, 300)

(2.858497199156411, 1.3149276685170943, 1.543569530639317, 0.2014551048135256)

MieScattering.mie_scattering_n — Function

mie_scattering_n(m, x, nmax::Int=-1)

Calculate scattering efficiency of a sphere

Arguments

m: relative refractive index ($n_\mathrm{material} / n_\mathrm{environment}$)

x: size parameter (wavenumber * radius)

namx: maximun order of resonances being considered (nmax = 1: dipole, 2: dipole+quadrupole, ..., -1: sufficiently large value determined by x)

Return values

Qsca: Scattering efficiency (total up to n-th order resonances)

MieScattering.mie_S1_S2 — Function

mie_S1_S2(m,x,mu)

Calculates $S_1$ and $S_2$ at $\mu = cos(\theta)$, where $\theta$ is the scattering angle.

Uses mie_ab to calculate a_n and b_n, and mie_pi_tau to calculate $\pi_n$ and $\tau_n$. $S_1$ and $S_2$ are calculated by:

\[S_1 = \sum\limits_{n=1}^{n_{max}} \frac{2n+1}{n(n+1)} (a_n \pi_n + b_n \tau_n)\]

\[S_2 = \sum\limits_{n=1}^{n_{max}} \frac{2n+1}{n(n+1)} (a_n \tau_n + b_n \pi_n)\]

Parameters

m: The complex refractive index with the convention m = n+ik.

x: The size parameter $x = \pi \, d_p / \lambda$.

mu: The cosine of the scattering angle.

Return values

S1, S2: The $S_1$ and $S_2$ values.

MieScattering.mie_S1_S2_n — Function

mie_S1_S2(m, x, mu, n::Int)

calculate n-th order S1,S2 component of scattering matrix.

MieScattering.scattering_function — Function

scattering_function(theta, m, x)

Creates arrays for plotting the angular scattering intensity functions in theta-space with parallel, perpendicular, and unpolarized light. Also includes an array of the angles for each step. This angle can be in either degrees, radians, or gradians for some reason. The angles can either be geometrical angle or the qR vector (see Sorensen, M. Q-space analysis of scattering by particles: a review. J. Quant. Spectrosc. Radiat. Transfer 2013, 131, 3-12 http://www.sciencedirect.com/science/article/pii/S0022407313000083). Uses mie_S1_S2 to compute $S_1$ and $S_2$, then computes parallel, perpendicular, and unpolarized intensities by

\[SL(\theta) = |S_1|^2\]

\[SR(\theta) = |S_2|^2\]

\[SU(\theta) = \frac{1}{2}(SR + SL)\]

Arguments

theta: An array of the angles used in calculations.

m: The complex refractive index with the convention $m = n+ik$.

wavelength: The wavelength of incident light, in nanometers.

diameter: The diameter of the particle, in nanometers.

n_medium: The refractive index of the surrounding medium. This must be positive, nonzero, and real. Any imaginary part will be discarded.

Return values

SL: An array of the scattered intensity of left-polarized (perpendicular) light. Same size as the theta array.

SR: An array of the scattered intensity of right-polarized (parallel) light. Same size as the theta array.

SU: An array of the scattered intensity of unpolarized light, which is the average of SL and SR. Same size as the theta array.

MieScattering.scattering_function_n — Function

scattering_function_n(theta, m, x)